Refactor meta display into component and add tooltip version
