+++ frontend/package.json
@@ -27,6 +27,7 @@
     "chart.js": "^2.1.6",
     "json-formatter-js": "^1.1.0",
     "moment": "^2.15.1",
+    "node-uuid": "^1.4.7",
     "react": "^15.2.1",
     "react-dom": "^15.2.1",
     "react-redux": "^4.4.5",
