+++ frontend/src/components/job/taskGroups.js
@@ -3,7 +3,7 @@ import {connect} from "react-redux";
 import {NomadLink} from "../link";
 import Table from "../table";
 import JSON from "../json";
-import { collectMeta } from '../../helpers/meta'
+import MetaDisplay from '../meta'
 
 class JobTaskGroups extends Component {
     render() {
@@ -23,7 +23,7 @@ class JobTaskGroups extends Component {
                     <td><NomadLink taskGroupId={taskGroup.ID} jobId={job.ID} short="true"/></td>
                     <td>{taskGroup.Name}</td>
                     <td>{taskGroup.Count}</td>
-                    <td>{collectMeta(taskGroup.Meta)}</td>
+                    <td><MetaDisplay metaBag={taskGroup.Meta} asTooltip={true} /></td>
                     <td>{taskGroup.RestartPolicy.Mode}</td>
                 </tr>
             )
